bacparser: add parser for 2012
Year 2012 is the back to the roots year: less crap, simpler HTML code.

The tr elements containing the results don't have the hint attribute
anymore, it's been replaced with onclick. Therefor move the old
xpath_get_trs method from BaseParser to Parser2006 and put a new version
in Parser2012.

The data model for the results is the same as in 2010.